:recycle: merge handler.go to client.go
